pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.6",
    "python-dotenv>=1.1.0",
]
